refactor: ReportCategory를 report feature 도메인으로 이동 + 버그 제보 TODO 명시 #233
