+++ lib/features/chat/presentation/widgets/chat_overlay.dart
@@ -5,7 +5,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 import '../../../report/presentation/providers/report_provider.dart';
-import '../../domain/constants/report_categories.dart';
+import '../../../report/domain/constants/report_categories.dart';
 import '../../../../core/constants/app_colors.dart';
 import '../../../../core/constants/spacing_and_radius.dart';
 import '../../../../core/constants/text_styles.dart';
@@ -144,20 +144,23 @@ class _ChatOverlayState extends ConsumerState<ChatOverlay> {
       onBlock: (participantId) {
         ref.read(chatNotifierProvider.notifier).blockUser(participantId);
       },
-      onReport: ({
-        required ReportCategory category,
-        required String messageContent,
-        required int reportedParticipantId,
-        String? etcReason,
-      }) async {
-        await ref.read(reportRepositoryProvider).reportChat(
-          gameId: widget.gameId,
-          reportedParticipantId: reportedParticipantId,
-          messageContent: messageContent,
-          category: category,
-          etcReason: etcReason,
-        );
-      },
+      onReport:
+          ({
+            required ReportCategory category,
+            required String messageContent,
+            required int reportedParticipantId,
+            String? etcReason,
+          }) async {
+            await ref
+                .read(reportRepositoryProvider)
+                .reportChat(
+                  gameId: widget.gameId,
+                  reportedParticipantId: reportedParticipantId,
+                  messageContent: messageContent,
+                  category: category,
+                  etcReason: etcReason,
+                );
+          },
     );
   }
