+++ lib/features/report/domain/constants/report_categories.dart
